improve constraint speed

+++ python/text_utils/inference/utils.py
@@ -167,7 +167,7 @@ def constraint_logit_fn(
                 zeros[i] = logits[i]
                 continue
 
-            indices = torch.from_numpy(constraint.get()).to(torch.long)
+            indices = torch.from_numpy(constraint.get())
             zeros[i, indices] = logits[i, indices]
 
             if constraint.is_match():
